config/types/files: add support for local files
This adds one more file contents source "local", in addition to
"inline" and "remote". It takes a filename, and includes the contents of
that file. This is useful when you don't want to pollute the ign file
with many large inline files, or when you don't want to indent the
contents of the file to fit the yaml syntax. Also, editing such a local
config file is easier when it is a separate file, because an editor can
enable format-specific features.

Local files are read from the local directory passed to --files-dir (to
be added in a later commit), so you could track both under source
control.