*: remove versioning from function names
for a while there was a thought that ct would provide the functionality
to convert between mutliple different specs, hence having versioned
functions for each different spec version. this never manifested itself,
and the effect was that everytime the spec bumped, we would have to
search and replace every function name with the new spec version,
constantly breaking api compatibility for no particular benefit.
removing the version numbers fixes this problem.